# By Fjölnir Ásgeirsson (69) and Ken Myers (1)
# Via Fjölnir Ásgeirsson (2) and Fjölnir Ásgeirsson (1)
* 'master' of https://github.com/fjolnir/DatabaseKit: (70 commits)
  Cached column names
  Cleanup
  Added narrow
  Fixed test compile error
  made BEGINSWITH smarter, it now makes use of indices when possible
  Made NSExpression capable of adding tokens to a query
  SELECT table.* now also returns model objects
  Made unions handle sorting properly
  Added SELECT DISTINCT
  Made predicates aware of the query they are generating for, this allows for more optimizations
  Cleanup
  Added a way to get types of columns (Not implemented for postgres yet, and I'll probably expand this to return more information such as default values etc to allow for added introspection)
  fast enum implementation broke when row caching was removed. Removing for now
  Cleanup
  Made dirtykeys update correctly
  Made order an enum
  Added support for nested transactions (Only with sqlite atm)
  Fixed a couple of bugs in predicate->sql translator
  Added block transactions
  Made destroy inert if not inserted
  ...

+++ DatabaseKit/Utilities/Base64Extensions.m
@@ -26,7 +26,7 @@ static const short _base64DecodingTable[256] = {
 - (NSData *)decodeBase64
 {
     const char *objPointer = [self cStringUsingEncoding:NSUTF8StringEncoding];
-    int intLength = strlen(objPointer);
+    size_t intLength = strlen(objPointer);
     int intCurrent;
     int i = 0, j = 0, k;
   
@@ -107,7 +107,7 @@ static const short _base64DecodingTable[256] = {
     char *strResult;
   
     // Get the Raw Data length and ensure we actually have data
-    int intLength = [self length];
+    NSUInteger intLength = [self length];
     if (intLength == 0) return nil;
   
     // Setup the String-based Result placeholder and pointer within that placeholder
